Added ability to reset tasks
Refactored youtube-dl updating and added youtube-dl update task
